chore(cli): refactor postinstall script, move some tests
