+++ packages/botonic-cli/scripts/postinstall.js
@@ -16,23 +16,22 @@ try {
   const { Telemetry } = require('../lib/analytics/telemetry')
   const telemetry = new Telemetry()
   telemetry.trackInstalledBotonic()
+  console.log('\n✨ Botonic was installed successfully.\n')
+  console.log(
+    'Create your first chatbot with:\n\x1b[1mbotonic new myBot\x1b[0m\n'
+  )
 } catch (e) {
-  //Some users don't have the right permissions to
-  //create dirs at instal time. We delay it until
-  //they run their first command.
+  const { getSystemInformation } = require('../lib/util/processes')
   if (process.env.BOTONIC_DISABLE_ANALYTICS !== '1') {
     const Analytics = require('analytics-node')
     var analytics = new Analytics('YD0jpJHNGW12uhLNbgB4wbdTRQ4Cy1Zu', {
       flushAt: 1,
     })
+    const properties = { ...getSystemInformation(), error: String(e) }
     analytics.track({
       event: 'Installed Botonic CLI',
       anonymousId: Math.round(Math.random() * 100000000),
+      properties,
     })
   }
 }
-
-console.log('\n✨ Botonic was installed successfully.\n')
-console.log(
-  'Create your first chatbot with:\n\x1b[1mbotonic new myBot\x1b[0m\n'
-)
